+++ CHANGELOG.md
@@ -1,4 +1,27 @@
 
+<a name="v3.1.1"></a>
+## [v3.1.1](https://github.com/aleyoscar/resurrexit/compare/3.1.0...v3.1.1) (2025-06-20)
+
+### Feat
+
+* **web:** Sort account custom lists by name
+* **web:** Updated service worker cache list
+
+### Fix
+
+* **web:** Caching not updating static files
+* **web:** Removed datalist & implemented custom autocomplete
+
+### Refactor
+
+* **web:** Removed unnecessary console.logs
+* **web:** Organized and commented code
+
+
+<a name="3.1.0"></a>
+## [3.1.0](https://github.com/aleyoscar/resurrexit/compare/v3.1.0...3.1.0) (2025-06-20)
+
+
 <a name="v3.1.0"></a>
 ## [v3.1.0](https://github.com/aleyoscar/resurrexit/compare/v3.0.0...v3.1.0) (2025-06-20)